requirements.txt
python-telegram-bot==20.5
pexpect==4.9.0
pyte==0.8.2
nest-asyncio==1.5.8
sqlalchemy==2.0.23
cryptography==41.0.7
python-decouple==3.8
aiosqlite==0.19.0